requirements.txt
Flask~=2.2.3
Pillow~=9.4.0
gunicorn~=20.1.0